Background: * Changed logo bar to be a table row, rather than image, for faster * loading. * Extracted peos logo from old background image. * Changed background to white, for readability and faster loading.
Function:
* Removed resource form from create process sequence.  This proved to
be too cumbersome; probably, resources should be bound on first
reference, rather than all at once up-front.
* Changed opening page to be a simple gateway, rather than login.
* Normalized look and feel so all pages have the same header and
background.

Style:
* Factored duplicate code into separate html.[ch] modules.
* Removed relative paths from includes.
* Modified tests to print one-line message on success.
*

+++ ui/web2/cgi-bin/test_main.sh
@@ -1,7 +1,8 @@
 #!/bin/sh
-echo -n "."
-echo TEST_MAIN.SH
-rm proc_table.dat
+echo -n "$0..."
+
+test -e proc_table.dat && rm proc_table.dat
+
 export QUERY_STRING="name=proc_table&pw=test"
 create_testtable
 main.cgi > output
@@ -49,5 +50,7 @@ then
   echo
 fi
 
-rm output
-rm proc_table.dat
+test -e output && rm output
+test -e proc_table.dat && rm proc_table.dat
+
+echo done
